Move the server models unit tests into mzutils too
